refactor(basicAlgorithms): split 161-line file into separate algorithm modules
apps/registry/app/job-similarity/utils/basicAlgorithms.js (161→2 lines)

Created algorithms/ module:
- index.js (4 lines) - Re-exports all algorithms
- mst.js (51 lines) - Minimum Spanning Tree (Kruskal's algorithm)
- knn.js (35 lines) - K-Nearest Neighbors algorithm
- threshold.js (27 lines) - Similarity Threshold algorithm
- adaptive.js (50 lines) - Adaptive Threshold algorithm

Updated basicAlgorithms.js to re-export from new module structure.

Benefits:
- Main file reduced from 161 to 2 lines (99% reduction)
- Each algorithm isolated in its own file
- Easier to test individual algorithms
- Clear separation of concerns
- Import paths remain backward compatible

All lints passing ✓
